cupy uses 32-bit values, not 8-bit

+++ parse_game.py
@@ -27,7 +27,7 @@ def read_games(fn):
 
 def bb2array(b, flip=False):
     #x = numpy.zeros(64, dtype=numpy.int8)
-    x = cp.zeros(64,dtype=cp.int8)
+    x = cp.zeros(64,dtype=cp.int32)
     for pos, piece in enumerate(b.pieces):
         if piece != 0:
             color = int(bool(b.occupied_co[chess.BLACK] & chess.BB_SQUARES[pos]))
